fix(review): Codex R2 — clean residual (uint256,uint256,uint256) TokenConfig in 2 evidence scripts + radar test fixture
The deployed v0.20.0 factory packs InitConfig.TokenConfig as (uint128 tier1Limit, uint128
tier2Limit, uint256 dailyLimit) (AAStarGlobalGuard.sol). l4-beta4-gasless.ts + beta3-weighted-sig.ts
local factory ABIs still had the stale (uint256,uint256,uint256) shape → wrong selector → would
revert on-chain (same class as the #118 production fix). Corrected to uint128/uint128/uint256;
also aligned the abi-signature.test.ts radar fixture + expected canonical strings for consistency.

+++ tests/regression/l4-beta4-gasless.ts
@@ -41,7 +41,7 @@ const FACTORY_ABI = [
           { name: 'approvedAlgIds', type: 'uint8[]' }, { name: 'minDailyLimit', type: 'uint256' },
           { name: 'initialTokens', type: 'address[]' },
           { name: 'initialTokenConfigs', type: 'tuple[]', components: [
-            { name: 'tier1Limit', type: 'uint256' }, { name: 'tier2Limit', type: 'uint256' }, { name: 'dailyLimit', type: 'uint256' }] }] }],
+            { name: 'tier1Limit', type: 'uint128' }, { name: 'tier2Limit', type: 'uint128' }, { name: 'dailyLimit', type: 'uint256' }] }] }],
       outputs: [{ type: 'address' }] },
     { type: 'function', name: 'createAccount', stateMutability: 'nonpayable',
       inputs: [{ name: 'owner', type: 'address' }, { name: 'salt', type: 'uint256' },
@@ -52,7 +52,7 @@ const FACTORY_ABI = [
           { name: 'approvedAlgIds', type: 'uint8[]' }, { name: 'minDailyLimit', type: 'uint256' },
           { name: 'initialTokens', type: 'address[]' },
           { name: 'initialTokenConfigs', type: 'tuple[]', components: [
-            { name: 'tier1Limit', type: 'uint256' }, { name: 'tier2Limit', type: 'uint256' }, { name: 'dailyLimit', type: 'uint256' }] }] }],
+            { name: 'tier1Limit', type: 'uint128' }, { name: 'tier2Limit', type: 'uint128' }, { name: 'dailyLimit', type: 'uint256' }] }] }],
       outputs: [{ type: 'address' }] },
 ] as const;
